Check google group membership with hasMember
An alternate take on #141 and #95 that's considerably simpler, and may
be more reliable.  This method can also handle nested group
memberships, according to Google's API docs:

https://developers.google.com/admin-sdk/directory/v1/reference/members/hasMember

However, the above message doesn't handle members who are not a part
of the domain. The API returns a 400 for that case (as well as for
others). As a fallback, when the API returns a 400, this change will
try using the `get` API which works as expected for members who aren't
a part of the domain.
https://developers.google.com/admin-sdk/directory/v1/reference/members/get

Note that nested members who are not a part of the domain will not be
correctly detected in this case.